+++ src/components/login/Login.ts
@@ -1,0 +1,7 @@
+import LoginForm from './LoginForm';
+
+const Login = {
+  Form: LoginForm,
+};
+
+export default Login;
