+++ src/components/molecules/index.ts
@@ -1,5 +1,3 @@
-// eslint-disable-next-line import/no-cycle
-export * from './form';
 export * from './dashboardChart';
 export * from './dashboardTable';
 export * from './table';
